+++ src/Core/Resources/ConditionalList.php
@@ -23,6 +23,7 @@ use Generator;
 use IteratorAggregate;
 use JsonSerializable;
 use LaravelJsonApi\Contracts\Resources\Skippable;
+use Traversable;
 use function iterator_to_array;
 
 class ConditionalList implements IteratorAggregate, JsonSerializable
@@ -76,7 +77,7 @@ class ConditionalList implements IteratorAggregate, JsonSerializable
     /**
      * @inheritDoc
      */
-    public function getIterator()
+    public function getIterator(): Traversable
     {
         return $this->cursor();
     }
@@ -84,7 +85,7 @@ class ConditionalList implements IteratorAggregate, JsonSerializable
     /**
      * @inheritDoc
      */
-    public function jsonSerialize()
+    public function jsonSerialize(): array
     {
         return $this->all();
     }