+++ CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file. This project adheres to
 [Semantic Versioning](http://semver.org/) and [this changelog format](http://keepachangelog.com/).
 
+## Unreleased
+
+### Added
+
+- The `LazyRelation` class now has an `all()` method to get the related resources for a to-many relation as an array.
+
+### Fixed
+
+- Previously the `LazyRelation` class would always throw an exception if iterated over. This has now been fixed.
+
 ## [2.0.0] - 2022-02-08
 
 ### Added